+++ docs-src/pages/UiAutocomplete.vue
@@ -255,6 +255,7 @@
                                     <ul>
                                         <li><code>suggestion</code>: (Object or String) - the current suggestion</li>
                                         <li><code>query</code>: (String) - the current value of the autocomplete input (what the user has typed)</li>
+                                        <li><code>defaultFilter</code>: (Function) - Keen UI's default filter function. You can call this with the suggestion and query to defer to the default fuzzy filter behavior. Remember to return its result from your custom filter function if you're using it.</li>
                                     </ul>
 
                                     <p>The function should return <code>true</code> if the suggestion matches the query or <code>false</code> otherwise.</p>