+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Add new method `focus()` to focus programmatically.
 - Add new prop `sort` for a custom sort function.
+- Pass default filter function as third parameter to `filter` prop.
 
 ### UiButton
 
@@ -131,6 +132,7 @@
 - Add new method `clearSelection()` to clear the current selection.
 - Add new method `focus()` to focus programmatically.
 - Add new prop `sort` for a custom sort function.
+- Pass default filter function as third parameter to `filter` prop.
 
 ### UiSlider